+++ pkg/target/teams/teams_test.go
@@ -7,6 +7,7 @@ import (
 	"time"
 
 	"github.com/kyverno/policy-reporter/pkg/report"
+	"github.com/kyverno/policy-reporter/pkg/target"
 	"github.com/kyverno/policy-reporter/pkg/target/teams"
 )
 
@@ -103,7 +104,7 @@ func Test_TeamsTarget(t *testing.T) {
 			}
 		}
 
-		client := teams.NewClient("http://hook.teams:80", "", []string{}, false, testClient{callback, 200})
+		client := teams.NewClient("Teams", "http://hook.teams:80", false, &target.Filter{}, testClient{callback, 200})
 		client.Send(completeResult)
 	})
 
@@ -133,7 +134,7 @@ func Test_TeamsTarget(t *testing.T) {
 			}
 		}
 
-		client := teams.NewClient("http://hook.teams:80", "", []string{}, false, testClient{callback, 200})
+		client := teams.NewClient("Teams", "http://hook.teams:80", false, &target.Filter{}, testClient{callback, 200})
 		client.Send(minimalResult)
 	})
 	t.Run("Send Minimal InfoResult", func(t *testing.T) {
@@ -150,7 +151,7 @@ func Test_TeamsTarget(t *testing.T) {
 			}
 		}
 
-		client := teams.NewClient("http://hook.teams:80", "", []string{}, false, testClient{callback, 200})
+		client := teams.NewClient("Teams", "http://hook.teams:80", false, &target.Filter{}, testClient{callback, 200})
 		client.Send(minimalInfoResult)
 	})
 	t.Run("Send Minimal ErrorResult", func(t *testing.T) {
@@ -167,7 +168,7 @@ func Test_TeamsTarget(t *testing.T) {
 			}
 		}
 
-		client := teams.NewClient("http://hook.teams:80", "", []string{}, false, testClient{callback, 200})
+		client := teams.NewClient("Teams", "http://hook.teams:80", false, &target.Filter{}, testClient{callback, 200})
 		client.Send(minimalErrorResult)
 	})
 	t.Run("Send Minimal Debug Result", func(t *testing.T) {
@@ -196,11 +197,11 @@ func Test_TeamsTarget(t *testing.T) {
 			}
 		}
 
-		client := teams.NewClient("http://hook.teams:80", "", []string{}, false, testClient{callback, 200})
+		client := teams.NewClient("Teams", "http://hook.teams:80", false, &target.Filter{}, testClient{callback, 200})
 		client.Send(minimalDebugResult)
 	})
 	t.Run("Name", func(t *testing.T) {
-		client := teams.NewClient("http://localhost:9200", "", []string{}, true, testClient{})
+		client := teams.NewClient("Teams", "http://localhost:9200", true, &target.Filter{}, testClient{})
 
 		if client.Name() != "Teams" {
 			t.Errorf("Unexpected Name %s", client.Name())