+++ pkg/target/ui/ui.go
@@ -74,14 +74,10 @@ func (e *client) Send(result *report.Result) {
 	helper.ProcessHTTPResponse(e.Name(), resp, err)
 }
 
-func (e *client) Name() string {
-	return "UI"
-}
-
 // NewClient creates a new loki.client to send Results to Elasticsearch
-func NewClient(host, minimumPriority string, sources []string, skipExistingOnStartup bool, httpClient httpClient) target.Client {
+func NewClient(name, host string, skipExistingOnStartup bool, filter *target.Filter, httpClient httpClient) target.Client {
 	return &client{
-		target.NewBaseClient(minimumPriority, sources, skipExistingOnStartup),
+		target.NewBaseClient(name, skipExistingOnStartup, filter),
 		host + "/api/push",
 		httpClient,
 	}